Almost finished conversion to scripts
