+++ module.json
@@ -9,15 +9,23 @@
     "system": [
         "pf2e"
     ],
+    "dependencies": [
+        {
+            "name": "lib-wrapper"
+        }
+    ],
     "url": "https://github.com/JDCalvert/FVTT-PF2e-Ranged-Combat",
     "manifest": "https://github.com/JDCalvert/FVTT-PF2e-Ranged-Combat/releases/latest/download/module.json",
     "download": "https://github.com/JDCalvert/FVTT-PF2e-Ranged-Combat/archive/1.3.0.zip",
     "readme": "https://github.com/JDCalvert/FVTT-PF2e-Ranged-Combat",
     "bugs": "https://github.com/JDCalvert/FVTT-PF2e-Ranged-Combat/issues",
     "changelog": "https://github.com/JDCalvert/FVTT-PF2e-Ranged-Combat/blob/main/CHANGELOG.md",
-    "scripts": [
+    "esmodules": [
         "scripts/config.js",
         "scripts/utils.js",
+        "scripts/dialog.js",
+        "scripts/actions/reload.js",
+        "scripts/actions/hunt-prey.js",
         "scripts/hunted-prey-hook.js",
         "scripts/fire-weapon-hook.js"
     ],
